requirements.txt
tb-mqtt-client==1.5
pyserial
mmh3